feedback: idx 최적화
