+++ src/main/kotlin/kr/hhplus/be/server/api/balance/controller/BalanceController.kt
@@ -9,6 +9,8 @@ import jakarta.validation.Valid
 import jakarta.validation.constraints.Positive
 import kr.hhplus.be.server.api.balance.dto.BalanceDto
 import kr.hhplus.be.server.api.balance.dto.request.ChargeBalanceRequest
+import kr.hhplus.be.server.api.balance.usecase.ChargeBalanceUseCase
+import kr.hhplus.be.server.api.balance.usecase.DeductBalanceUseCase
 import kr.hhplus.be.server.domain.balance.service.BalanceService
 import kr.hhplus.be.server.global.response.CommonApiResponse
 import org.springframework.http.ResponseEntity
@@ -25,7 +27,9 @@ import org.springframework.web.bind.annotation.RestController
 @Validated
 @Tag(name = "Balance", description = "잔액 관리 API")
 class BalanceController(
-    private val balanceService: BalanceService
+    private val balanceService: BalanceService,
+    private val chargeBalanceUseCase: ChargeBalanceUseCase,
+    private val deductBalanceUseCase: DeductBalanceUseCase
 ) {
 
 
@@ -33,7 +37,7 @@ class BalanceController(
     fun chargeBalance(
         @Valid @RequestBody request: ChargeBalanceRequest
     ): ResponseEntity<CommonApiResponse<BalanceDto.ChargeResult>> {
-        val point = balanceService.chargeBalance(request.userId, request.amount)
+        val point = chargeBalanceUseCase.execute(request.userId, request.amount)
         return ResponseEntity.ok(
             CommonApiResponse.Companion.success(
                 data = BalanceDto.ChargeResult.from(point, request.amount),
