+++ src/main/kotlin/kr/hhplus/be/server/domain/concert/models/Seat.kt
@@ -12,10 +12,8 @@ import java.time.LocalDateTime
 @Table(
     name = "seat",
     indexes = [
-        Index(name = "idx_seat_schedule_id_status_code", columnList = "schedule_id, status_code"),
-        Index(name = "idx_seat_schedule_id_seat_number", columnList = "schedule_id, seat_number"),
         Index(name = "idx_seat_schedule_status_number", columnList = "schedule_id, status_code, seat_number"),
-        Index(name = "idx_seat_schedule_id_price", columnList = "schedule_id, price")
+        Index(name = "idx_seat_schedule_price", columnList = "schedule_id, price")
     ]
 )
 data class Seat(
