+++ src/main/kotlin/kr/hhplus/be/server/domain/payment/models/Payment.kt
@@ -14,11 +14,8 @@ import java.time.LocalDateTime
 @Table(
     name = "payment",
     indexes = [
-        Index(name = "idx_payment_user_id", columnList = "user_id"),
-        Index(name = "idx_payment_status_code", columnList = "status_code"),
-        Index(name = "idx_payment_paid_at", columnList = "paid_at"),
-        Index(name = "idx_payment_user_id_status_code", columnList = "user_id, status_code"),
-        Index(name = "idx_payment_amount", columnList = "amount")
+        Index(name = "idx_payment_user_status", columnList = "user_id, status_code"),
+        Index(name = "idx_payment_paid_at", columnList = "paid_at")
     ]
 )
 data class Payment(
